bot/requirements.txt
asyncio
discord
DiscordUtils
PyNaCl
python-dotenv
yt-dlp
requests